Create GESTURE CONTROL CAR FINAL.pdf
Another project was based on the combination of Sensor Interfacing and basic knowledge of the Basic of Electronics Engineering. In that I made Gesture Control Car. And all the technical details are mentioned in the report. Also am giving the video link : https://youtu.be/Fy6HRAjAhog
so that you can find how the model actually runs.